Removed body frame from metaframe WireFormatVersion is uint

+++ src/kino.Messaging/Versioning.cs
@@ -2,7 +2,7 @@
 {
     public class Versioning
     {
-        public const int WireFormatV5 = 5;
-        public const int WireFormatV6 = 6;
+        public const ushort WireFormatV5 = 5;
+        public const ushort WireFormatV6 = 6;
     }
 }
